backend/pyproject.toml
[project]
name = "uv-fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.6.1",
    "autogen-ext[openai]>=0.6.1",
    "azure-identity>=1.15.0",
    "fastapi[standard]>=0.112.1",
    "pydantic-settings>=2.9.1",
    "ruff>=0.11.13",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[dependency-groups]
test = [
    "pytest>=8.4.0",
]
